Code Revamp
